requirements.txt
# 博客爬虫项目依赖
# 支持多种文档框架：ReadTheDocs, RBook, MkDocs, Sphinx, Teadocs, Docsify

# Scrapy爬虫框架
scrapy>=2.11.0

# HTML到Markdown转换
markdownify>=0.11.0

# HTML解析
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Item适配器
itemadapter>=0.8.0

# YAML解析（用于metadata）
pyyaml>=6.0
